Use new ActiveSupport::Callbacks::Chain
In Rails 4.1 the chain class wraps an object @chain, which stores the
information about the callbacks. We use this class for deleting before and after
filters that should be called only for specific submodules. Therefore we need to
check for the Rails version and decide which way to use.

Anyway, this way of removing callbacks is not very pretty. Let's look for
something better soon.