Store user_id in session as string
Currently we store in session `user.id` which in some cases is integer, but in
some cases (like Mongoid) is an object.
What's the problem with storing object? First thing, it might break
compatibility between different ORM versions (Mongoid changed class of this
object from v2 to v3). Second, the bigger issue, is that new serializer in Rails
uses JSON instead of YAML, so it can store only selected types, and it parses
custom objects incorrectly, which may break the authentication feature.

This commit changes the way we store user id to plain string. This will make it
being stored correctly no matter what serializer is used.